requirements.txt
agent-development-kit
google-cloud-aiplatform